Move permission checks for user requests to helpers and update views and tests
